+++ CHANGES.md
@@ -6,6 +6,7 @@ Changelog
 
 * Remove all views on `reset` event
 * Debounce `sortAndPositionAllItems` on `add` event
+* New property for Backbone.OrderedListView: `sortImmediatelyOnAdd`
 
 1.0.2 (2018-01-29)
 ------------------